move report functions to their respective modules inside golem.report package
